move relevant files to /collect folder
