+++ src/components/[guild]/collect/components/RequirementsCard.tsx
@@ -1,10 +1,10 @@
 import { Skeleton, Stack, Text, useColorModeValue } from "@chakra-ui/react"
-import Card from "components/common/Card"
 import LogicDivider from "components/[guild]/LogicDivider"
-import CollectNftButton from "components/[guild]/Requirements/components/GuildCheckout/components/buttons/CollectNftButton"
 import SwitchNetworkButton from "components/[guild]/Requirements/components/GuildCheckout/components/buttons/SwitchNetworkButton"
-import { useCollectNftContext } from "components/[guild]/Requirements/components/GuildCheckout/components/CollectNftContext"
 import RequirementDisplayComponent from "components/[guild]/Requirements/components/RequirementDisplayComponent"
+import CollectNftButton from "components/[guild]/collect/components/CollectNftButton"
+import { useCollectNftContext } from "components/[guild]/collect/components/CollectNftContext"
+import Card from "components/common/Card"
 import { Chains } from "connectors"
 import { Fragment } from "react"
 import { Logic, Requirement } from "types"
