+++ src/components/[guild]/collect/components/ShareButton.tsx
@@ -6,7 +6,7 @@ import {
   MenuList,
   useClipboard,
 } from "@chakra-ui/react"
-import { useCollectNftContext } from "components/[guild]/Requirements/components/GuildCheckout/components/CollectNftContext"
+import { useCollectNftContext } from "components/[guild]/collect/components/CollectNftContext"
 import { CopySimple, ShareNetwork, TwitterLogo } from "phosphor-react"
 import useNftDetails from "../hooks/useNftDetails"
