+++ src/pages/[guild]/collect/[chain]/[address].tsx
@@ -8,6 +8,7 @@ import {
   useBreakpointValue,
 } from "@chakra-ui/react"
 import CollectibleImage from "components/[guild]/collect/components/CollectibleImage"
+import { CollectNftProvider } from "components/[guild]/collect/components/CollectNftContext"
 import Details from "components/[guild]/collect/components/Details"
 import Links from "components/[guild]/collect/components/Links"
 import NftByRole from "components/[guild]/collect/components/NftByRole"
@@ -17,7 +18,6 @@ import ShareButton from "components/[guild]/collect/components/ShareButton"
 import TopCollectors from "components/[guild]/collect/components/TopCollectors"
 import useNftDetails from "components/[guild]/collect/hooks/useNftDetails"
 import useGuild from "components/[guild]/hooks/useGuild"
-import { CollectNftProvider } from "components/[guild]/Requirements/components/GuildCheckout/components/CollectNftContext"
 import { ThemeProvider, useThemeContext } from "components/[guild]/ThemeContext"
 import CardMotionWrapper from "components/common/CardMotionWrapper"
 import GuildLogo from "components/common/GuildLogo"
