+++ src/platforms/ContractCall/CollectNftModalButton.tsx
@@ -17,13 +17,13 @@ import { usePostHogContext } from "components/_app/PostHogProvider"
 import Button from "components/common/Button"
 import { Modal } from "components/common/Modal"
 import { Chains } from "connectors"
-import AlphaTag from "./components/AlphaTag"
-import { useCollectNftContext } from "./components/CollectNftContext"
-import { useGuildCheckoutContext } from "./components/GuildCheckoutContex"
-import TransactionStatusModal from "./components/TransactionStatusModal"
-import OpenseaLink from "./components/TransactionStatusModal/components/OpenseaLink"
-import CollectNftButton from "./components/buttons/CollectNftButton"
-import SwitchNetworkButton from "./components/buttons/SwitchNetworkButton"
+import AlphaTag from "../../components/[guild]/Requirements/components/GuildCheckout/components/AlphaTag"
+import { useGuildCheckoutContext } from "../../components/[guild]/Requirements/components/GuildCheckout/components/GuildCheckoutContex"
+import TransactionStatusModal from "../../components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusModal"
+import OpenseaLink from "../../components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusModal/components/OpenseaLink"
+import SwitchNetworkButton from "../../components/[guild]/Requirements/components/GuildCheckout/components/buttons/SwitchNetworkButton"
+import CollectNftButton from "../../components/[guild]/collect/components/CollectNftButton"
+import { useCollectNftContext } from "../../components/[guild]/collect/components/CollectNftContext"
 
 const CollectNftModalButton = () => {
   const { captureEvent } = usePostHogContext()
